pyproject.toml
[tool.poetry]
name = "sphiwe"
version = "0.1.0"
description = ""
authors = ["Nazire Mathe <nazireemathe@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
django = "4.2.11"
django-split-settings = "^1.3.1"
dotmap = "^1.3.30"

whitenoise = "^6.6.0"
requests = "^2.32.3"
django-crispy-forms = "^2.2"
crispy-bootstrap4 = "^2024.1"
channels = "^4.1.0"
qrcode = "^7.4.2"
pydub = "^0.25.1"
django-allauth = "^65.1.0"
jwt = "^1.3.1"
pyjwt = "^2.9.0"
python-barcode = "^0.15.1"
python-dateutil = "^2.9.0.post0"
psycopg2-binary = "^2.9.10"
openai = "0.28"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
